# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "List all reservations",
                "operationId": "1fea0748719a5cbced2b611e67240778",
                "responses": {
                    "200": {
                        "description": "List of reservations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a new reservation",
                "operationId": "8a099249c009bb79e91d5ded3b235096",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "required": [
                                            "attributes",
                                            "relationships"
                                        ],
                                        "properties": {
                                            "attributes": {
                                                "required": [
                                                    "date",
                                                    "time",
                                                    "number_of_people"
                                                ],
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "time": {
                                                        "type": "string",
                                                        "example": "18:30"
                                                    },
                                                    "number_of_people": {
                                                        "type": "integer",
                                                        "minimum": 1
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "relationships": {
                                                "required": [
                                                    "client",
                                                    "table"
                                                ],
                                                "properties": {
                                                    "client": {
                                                        "properties": {
                                                            "data": {
                                                                "required": [
                                                                    "id"
                                                                ],
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "table": {
                                                        "properties": {
                                                            "data": {
                                                                "required": [
                                                                    "id"
                                                                ],
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully"
                    }
                }
            }
        },
        "/api/v1/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get reservation by ID",
                "operationId": "bd8d4153599117084526a3536d3c2b3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation data"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update a reservation",
                "operationId": "2ae6de381451f2ee52a4700b5f191ee0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "required": [
                                            "attributes",
                                            "relationships"
                                        ],
                                        "properties": {
                                            "attributes": {
                                                "required": [
                                                    "date",
                                                    "time",
                                                    "number_of_people"
                                                ],
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "time": {
                                                        "type": "string",
                                                        "example": "18:30"
                                                    },
                                                    "number_of_people": {
                                                        "type": "integer",
                                                        "minimum": 1
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "relationships": {
                                                "required": [
                                                    "client",
                                                    "table"
                                                ],
                                                "properties": {
                                                    "client": {
                                                        "properties": {
                                                            "data": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "table": {
                                                        "properties": {
                                                            "data": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Delete a reservation",
                "operationId": "f2f0d865144c44afb03db5f89c4b197c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reservation deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}